+++ src/app/services/list/abstract-list.service.ts
@@ -1,4 +1,4 @@
-import Movie from '../model/movie';
+import Movie from '../../model/movie';
 
 export abstract class AbstractListService {
