+++ src/app/services/movie/movie.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 
 import { MovieService } from './movie.service';
-import { MovieRepository } from '../repositories/movie-repository.service';
+import { MovieRepository } from '../../repositories/movie-repository.service';
 
 describe('MovieService', () => {
   let service: MovieService;
