+++ src/app/services/movie/movie.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { MovieRepository } from '../repositories/movie-repository.service';
-import Movie from '../model/movie';
+import { MovieRepository } from '../../repositories/movie-repository.service';
+import Movie from '../../model/movie';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import Cast from '../model/cast';
+import Cast from '../../model/cast';
 
 @Injectable({
   providedIn: 'root'
