+++ src/app/services/watch-list/watch-list.service.spec.ts
