+++ src/app/views/movies/movies.component.spec.ts
@@ -4,7 +4,7 @@ import { MoviesComponent } from './movies.component';
 import { NO_ERRORS_SCHEMA } from '@angular/compiler';
 import { of } from 'rxjs';
 import movies from 'src/app/mock/movies-mock';
-import { MovieService } from 'src/app/services/movie.service';
+import { MovieService } from 'src/app/services/movie/movie.service';
 
 describe('MoviesComponent', () => {
   let component: MoviesComponent;
