+++ src/app/views/seen-movies/seen-movies.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SeenMoviesComponent } from './seen-movies.component';
-import { SeenListService } from 'src/app/services/seen-list.service';
+import { SeenListService } from 'src/app/services/seen-list/seen-list.service';
 import { of } from 'rxjs';
 import movies from 'src/app/mock/movies-mock';
