+++ src/app/views/settings/settings.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SettingsComponent } from './settings.component';
-import { ApiKeyService } from 'src/app/services/api-key.service';
+import { ApiKeyService } from 'src/app/services/api-key/api-key.service';
 
 describe('SettingsComponent', () => {
   let component: SettingsComponent;
